feat: implement generate_for_json method
Before
======

There is a `genereate_for_jed`. This defines the translations in a JSON
hash and assigns it to a variable.

After
=====

Implemented `generate_for_json`, which generates the same json hash as
`generate_for_jed`, but does not add all the javascript boilerplate.
This is why the code is mostly deletes.